pyproject.toml
[project]
name = "mood-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "polars>=1.35.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "typer>=0.20.0",
]
